Implemented own authentication provider.
This new authentication provider for simple prompts (username and password) is basically the same as Subversions original, but I added encryption for the data on Win2k and later.
Also, the auth data is now stored in the registry under HKCU which makes it also harder for other users to even read the encrypted data.
The encryption feature is only available on Win2k and later because the API funtion is not available on earlier systems. That's also the reason why those API functions are explicitely linked (loaded dynamically) so the program won't crash on OS's where these API functions are not available.

If the encryption API's are not available, then the default function is used, which is the same as Subversion uses.

git-svn-id: svn://svn.code.sf.net/p/tortoisesvn/code/trunk@1459 154be670-7969-47af-9ab0-c6fb378d2295