Patch from Simon Large (slarge@blazepoint.co.uk):
This implements the colour interpolation scheme described on the user list. I have tested it with quite a few of the Win2K colour schemes and it seems to work in most cases. It does not work that well with the reverse colour high-contrast schemes because there is not a big difference between green-on-black and green-on-very-dark-grey. But there again, the scintilla bit ignores the colour scheme altogether.

I have also attached the makefile in case anyone else wants to have a go at building TBlame the hard way.

git-svn-id: svn://svn.code.sf.net/p/tortoisesvn/code@1701 154be670-7969-47af-9ab0-c6fb378d2295