Patch from Damian Powell (damian@shadow-angel.com):
After the CSVNStatusListCtrl has handled a context menu command, at checks
to see if the number of items has changed (perhaps an item was reverted,
deleted, ignored or otherwise removed from the list for some reason) and if
it has, a registered message is sent to the parent window. The CLogPromptDlg
class handles this message to do a check for an empty list. If the list is
empty, then the "There's nothing for TortoiseSVN to do here..." message is
displayed and the dialog closes.

Also, after sending one of my earlier patches, I realised that my
fixed_array<T> class was basically a poor cousin to std::vector<T,A> so I
enhanced it so that it is now a full subclass of the same name and template
parameters within the stdex namespace. It doesn't have all of the STL
vector's constructors yet but they would be easy enough to add if they are
required.

git-svn-id: svn://svn.code.sf.net/p/tortoisesvn/code@2016 154be670-7969-47af-9ab0-c6fb378d2295