Patch from Will Dean (svn@indcomp.co.uk): The attached patch addresses some severe performance problems with TMerge on very large files.
The main problem was that someone had misguidedly set the growth behaviour of the CStringArrays which hold the file to be a linear growth of 10 lines at a time.  This is really not suitable at all for a large file - an exponential growth is much better.  The default behaviour of CStringArray (and the other MFC arrays, which all template off the same code) *is* exponential growth, but it's strongly biased towards minimising memory waste and away from minimising the number of allocations/copies.

So even though the CStringArray classes did better at their default settings than with the 'grow-by-ten' setting, they still weren't very good.

To avoid all this lousy behaviour but not have to re-write vast scads of TMerge, I have written a class template which makes a STL vector look like an MFC array.  I have replaced all the CStringArray and CDWORDArrays with my class.

I have also stopped each file being loaded twice from disk rather than just being copied in memory.

The upshot of all this is that I can diff two 160000-line files together perfectly reasonably.  The previous version took a lot longer to load.

git-svn-id: svn://svn.code.sf.net/p/tortoisesvn/code@2144 154be670-7969-47af-9ab0-c6fb378d2295