Show only those properties in the combobox which actually can be used on the selected item.
I.e. if a file is selected in the Subversion properties tab, then only add those properties which can be used on files, and if a folder is selected only show those which can be used on folders.
If more than one item (file or folder) is selected, all properties are added to the combobox.

git-svn-id: svn://svn.code.sf.net/p/tortoisesvn/code/trunk@2627 154be670-7969-47af-9ab0-c6fb378d2295