Restructure the DUG. Issue #16
This is an attempt to make the DUG clearer and easier to use.

- Rework 'Handling TSVN' and 'Authentication' into a new 'Getting Started'.
- Rename 'Get/Set Properties of Files and Folders' to 'Project Properties'.
- Add a picture of the right drag menu.
- Reorder the sections closer to the order they are likely to be encountered.

Note: The German files have been modified to show the new structure,
but some translations are needed in:
- de/dug_general.xml
- de/dug_propertypage.xml

RightDragMenu.png needs to be recaptured in XP, and in German.

The German docs do not yet have a RevGraph.xml file.

git-svn-id: svn://svn.code.sf.net/p/tortoisesvn/code/trunk@2802 154be670-7969-47af-9ab0-c6fb378d2295