patch by Aras Pranckevièius <nearaz_at_gmail.com>: http://svn.haxx.se/tsvn/archive-2008-10/0047.shtml
Log statistics dialog used to disable pie graph type based on number 
of weeks, even if after large amount of weeks the graph turns into 
months and so on. This was somewhat confusing, as after long log 
history the pie graph would be disabled for no apparent reason. 

The proper thing would be at least using GetUnitCount() instead of 
m_nWeeks in CStatGraphDlg::OnCbnSelchangeGraphcombo(). However, I 
think what makes most sense is just to enable the pie graph always. 
The worst that can happen is that it will have 14 pies for weeks (at 
15 weeks it turns into months already). The attached patch does that.



git-svn-id: svn://svn.code.sf.net/p/tortoisesvn/code/trunk@14211 154be670-7969-47af-9ab0-c6fb378d2295