Properly separate the different lines we use in TortoiseMerge: Screen lines : the lines shown on the screen View lines   : the lines in the diff view, not all of them are necessarily shown (collapsed sections)
To do that, introduce a new vector member that's used to convert screen lines to view lines. That vector is re-calculated every time the view data changes.
Also rename some methods to make it clear that they're working on view lines, the default for methods is to work with screen lines.

This also fixes some annoying scrolling problems when unmodified sections are collapsed.

git-svn-id: svn://svn.code.sf.net/p/tortoisesvn/code/trunk@19905 154be670-7969-47af-9ab0-c6fb378d2295