Message from the future: this commit is reverted in r21606.
Don't mark the range below the last reported revision
as "can be skipped". That information would be used
very rarely in the first place.

This is a workaround for SVN server implementations
that don't report the initial "add" of the given path
in a "svn log $path --stop-on-copy" call (e.g. GoogleCode). 

git-svn-id: svn://svn.code.sf.net/p/tortoisesvn/code/trunk@21587 154be670-7969-47af-9ab0-c6fb378d2295